+++ cli.js
@@ -139,22 +139,22 @@ function mri (args, opts) {
 
 const list_columns = {
   domain: 'DOMAIN',
-  template: 'TEMPLATE',
   state: 'STATE',
   address: 'PROXY ADDRESS',
   certificate: 'CERTIFICATE',
   keepalive: 'KEEPALIVE',
-  additional: 'ADDITIONAL DOMAINS'
+  additional: 'ADDITIONAL DOMAINS',
+  template: 'TEMPLATE',
 };
 
 const length = {
   domain: 6,
-  template: 8,
   state: 6,
   address: 13,
   certificate: 11,
   keepalive: 9,
-  additional: 18
+  additional: 18,
+  template: 8,
 };
 
 const pad   = 4;
@@ -173,7 +173,7 @@ const list_table = function ( proxies ) {
         proxy[column] = proxy[column].join(', ');
 
       // Convert column value to string, in case it's a number
-      proxy[column] = '' + proxy[column];
+      proxy[column] = '' + (proxy[column] || '');
 
       // Replace certificate boolean
       if ( column == 'certificate' )
@@ -227,6 +227,14 @@ function parseUrl(input) {
 function titleCase(input) {
   input = input.trim().toLowerCase();
   return input.charAt(0).toUpperCase() + input.slice(1);
+}
+
+function isValidTimestamp(value) {
+  if (!Number.isInteger(value)) return false;
+
+  const date = new Date(value);
+  
+  return !isNaN(date.getTime()) && date.getTime() === value;
 }
 
 class StandardTemplate {
@@ -66244,19 +66252,28 @@ class LibSSL {
     }
     return readFileSync(this.path.domain_key, 'utf8');
   }
-  
-  // Check certificate expiration
-  expiring() {
-    if (!existsSync(this.path.cert)) return { expired: true };
 
+  // Certificate exists
+  certExists() {
+    return existsSync(this.path.cert);
+  }
+
+  // Certificate date
+  certDate() {
     const certPem = readFileSync(this.path.cert, 'utf8');
     const cert = acme.crypto.readCertificateInfo(certPem);
     const remaining_days = (new Date(cert.notAfter) - new Date()) / (1000 * 60 * 60 * 24);
 
     return {
+      certDate: cert.notAfter,
       days: remaining_days,
-      expired: remaining_days < this.expire_days
-    };
+      expired: certDate.days < this.expire_days
+    }
+  }
+  
+  // Check certificate expiration
+  expiring() {
+    return !this.certExists() ? { expired: true } : this.certDate();
   }
 
   // Asynchronous function to handle the HTTP challenge
@@ -66319,8 +66336,8 @@ class LibSSL {
     writeFileSync(this.path.csr, certificateCsr);
     writeFileSync(this.path.domain_key, certificateKey);
 
-    const order = await client.createOrder({  
-      identifiers: Array.from(new Set([this.proxy.domain].concat(this.proxy.additional)))  
+    const order = await client.createOrder({
+      identifiers: Array.from(new Set([this.proxy.domain].concat(this.proxy.additional)))
         .map(domain => ({ type: 'dns', value: domain }))  
     });
 
@@ -66358,7 +66375,10 @@ class LibSSL {
     writeFileSync(this.path.cert, cert);
     writeFileSync(this.path.fullchain, cert);
 
-    return { success: `Successfully generated certificate for ${this.proxy.domain}` };
+    // Get the certificate expiration date
+    const certData = acme.crypto.readCertificateInfo(cert);
+
+    return { success: `Successfully generated certificate for ${this.proxy.domain}`, certDate: certData.notAfter };
   };
 
   cleanup(wellknown) {
@@ -66367,8 +66387,8 @@ class LibSSL {
 
   // Main function to check, generate and renew certificate
   async generate() {
-    const { days, expired } = this.expiring();
-    return expired ? this.order() : { success: `Certificate for ${this.proxy.domain} still valid for ${Math.floor(days)} days` };
+    const { days, certDate, expired } = this.expiring();
+    return expired ? this.order() : { success: `Certificate for ${this.proxy.domain} still valid for ${Math.floor(days)} days`, certDate };
   }
 
 }
@@ -66417,9 +66437,6 @@ class Narnia {
   }
 
   list() {
-    if ( !existsSync(this.config.proxy_dir) )
-      return { error: `Proxy directory ${this.config.proxy_dir} doesn't exist\nRun 'narnia install' to create it` }
-
     this.retrieve();
     list_table(this.proxies);
   }
@@ -66581,19 +66598,18 @@ class Narnia {
     );
   }
 
-  async ssl( options ) {
-    const { proxy, path, error } = this.ensure(options, 'ssl');
+  ssl_generate( options, command = 'ssl:generate' ) {
+    let { proxy, path, error } = this.ensure(options, command);
     if ( error ) return { error };
 
-    if ( options.generate ) {
-      const staging = !!options?.staging;
-      return this.ssl_generate(path, proxy, staging)
-    }
-
-    return { error: `Missing or invalid option for 'narnia ssl ${proxy.domain}'` }
+    return this.ssl_single( proxy, path );
   }
 
-  async ssl_generate( path, proxy, staging ) {
+  async ssl_single( proxy, path ) {
+    if (!path) path = this.config.proxy_dir + proxy.domain;
+
+    const staging = !!options?.staging;
+
     // Enable .well-known directory and reload nginx
     let wellknown = true;
     this.save( path, proxy, wellknown );
@@ -66605,8 +66621,9 @@ class Narnia {
 
     const result = await libssl.generate( staging );
 
-    if ( !staging && result.success )
-      proxy.certificate = true;
+    if ( !staging && result.success ) {
+      proxy.certificate = result.certDate.getTime();
+    }
 
     // Disable .well-known directory and reload nginx
     wellknown = false;
@@ -66614,6 +66631,87 @@ class Narnia {
     this.reload();
 
     return result;
+  }
+
+  async ssl_renew ( options ) {
+    // 1. Single proxy renew
+
+    if (options.name)
+      return this.ssl_generate(options, 'ssl:renew')
+
+    // 2. All proxies renew
+    
+    // Retrieve and populate this.proxies
+    this.retrieve();
+
+    // Renew each one sequentially, to avoid being blocked
+    // by Let's Encrypt servers
+    for ( const proxy in this.proxies ) {
+      // Skip if there is no certificate
+      if (proxy.certificate == false) {
+        console.log( `[Skip]: Proxy ${domain} with SSL not enabled` );
+        continue;
+      }
+
+      // Ask for renew if there is a certificate
+      const response = await this.ssl_single(proxy);
+      this.print_response(response);
+    }
+
+    return { success: 'SSL renew operation concluded' }
+  }
+
+  async ssl_check() {
+    // Retrieve and populate this.proxies
+    this.retrieve();
+
+    // Iterate, ensuring the certificate expiration date is added when missing
+    // for retro-compatibility
+    Object.entries(this.proxies).forEach(this.ensure_ssl_date);
+
+    return { success: 'SSL check operation concluded' }
+  }
+
+  ensure_ssl_date( domain ) {
+    const proxy = this.proxies[domain];
+
+    // Skip if there is no certificate
+    if (proxy.certificate == false) {
+      console.log( `[Skip]: Proxy ${domain} with SSL not enabled` );
+      return;
+    }
+
+    // Skip if timestamp is already there
+    if (isValidTimestamp(proxy.certificate)) {
+      console.log( `[OK]: Proxy ${domain} already with certificate expiration` );
+      return;
+    }
+
+    // If there is a certificate but not a valid timestamp,
+    // this is probably a proxy from an old version of narnia
+    // Let's update it!
+
+    const libssl = new LibSSL(this.config, proxy);
+    const path  = this.config.proxy_dir + domain;
+
+    // If there is no certificate file, well, then we consider this proxy as
+    // certificate-less as a safe measure
+    if (!libssl.certExists()) {
+      proxy.certificate = false;
+      console.log( `[Change]: No certificate found on proxy ${domain}` );
+    }
+    // Otherwise, victory! The proxy now has a valid timestamp and can be
+    // updated!
+    else {
+      proxy.certificate = libssl.certDate().certDate.getTime();
+      console.log( `[Change]: Proxy ${domain} certificate expiration checked` );
+    }
+
+    // Save the changes
+    const wellknown = false;
+    this.save( path, proxy, wellknown );
+    
+    return;
   }
 
   reload() {
@@ -66647,6 +66745,23 @@ class Narnia {
     if ( !existsSync(path) ) return { error: `Proxy ${options.name} doesn't exist` }
 
     return { proxy: JSON.parse(readFileSync(path, 'utf8')), path }
+  }
+
+  ensure_config() {
+    let response = { error: null };
+
+    if ( !existsSync(this.config.proxy_dir) )
+      response.error = `Proxy directory ${this.config.proxy_dir} doesn't exist\nRun 'narnia install' to create it`;
+
+    return response;
+  }
+
+  print_response(response) {
+    if ( response?.error )
+      console.log( 'Error: ' + response.error );
+
+    if ( response?.success )
+      console.log( response.success );
   }
 
 }
@@ -66702,19 +66817,19 @@ const command = mri( process.argv.slice( 2 ), {
 });
 
 if ( command.help || ( process.argv.length <= 2 && process.stdin.isTTY ) ) {
-  console.log( 'Narnia version ' + '0.2.2' );
+  console.log( 'Narnia version ' + '0.4.0' );
   console.log( 'Narnia proxy manager help text go here' );
   process.exit();
 }
 
 if ( command.version ) {
-  console.log( 'Narnia version ' + '0.2.2' );
+  console.log( 'Narnia version ' + '0.4.0' );
   process.exit();
 }
 
 const config = readConf();
 const narnia = new Narnia( config );
-const mode   = command[ '_' ][ 0 ];
+const mode   = command[ '_' ][ 0 ].replace(':', '_');
 
 if ( typeof narnia[mode] != 'function' ) {
   console.log( `Invalid command 'narnia ${mode}'` );
@@ -66725,10 +66840,15 @@ if ( typeof narnia[mode] != 'function' ) {
 if ( command[ '_' ].length > 1 )
   command.name = command[ '_' ][ 1 ];
 
-const response = await narnia[mode](command);
+// Ensure installation and config directory
+let response = await narnia.ensure_config();
 
-if ( response?.error )
+if ( response?.error ) {
   console.log( 'Error: ' + response.error );
+  process.exit();
+}
 
-if ( response?.success )
-  console.log( response.success );
+// Call command
+response = await narnia[mode](command);
+narnia.print_response(response);
+process.exit();